[#18] :: Add move stack functionality and undo_move method
