+++ src/models/game/mod.rs
@@ -1,3 +1,4 @@
 pub mod block;
 pub mod board;
+pub mod move_;
 pub mod utils;
